refactor: create controller (https://github.com/Taiki130/ecsexec/issues/36)
