--- pkg/cli/region.go
@@ -1,17 +1,0 @@
-package cli
-
-import (
-	"github.com/manifoldco/promptui"
-)
-
-func promptRegion() (string, error) {
-	l := "Enter Region"
-	prompt := promptui.Prompt{
-		Label: l,
-	}
-	result, err := prompt.Run()
-	if err != nil {
-		return "", err
-	}
-	return result, nil
-}
