--- pkg/cli/session.go
@@ -1,37 +1,0 @@
-package cli
-
-import (
-	"encoding/json"
-	"fmt"
-	"os"
-	"os/exec"
-
-	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/service/ecs/types"
-	"github.com/aws/aws-sdk-go-v2/service/ssm"
-)
-
-func startSession(sess *types.Session, region string, target string) error {
-	sessJSON, _ := json.Marshal(sess)
-	endpoint := getSSMEndpoint(region)
-	payload := ssm.StartSessionInput{
-		Target: aws.String(target),
-	}
-	payloadJSON, err := json.Marshal(payload)
-	if err != nil {
-		return err
-	}
-
-	cmd := exec.Command(
-		"session-manager-plugin", string(sessJSON), region, "StartSession", "", string(payloadJSON), endpoint,
-	)
-	cmd.Stdout = os.Stdout
-	cmd.Stdin = os.Stdin
-	cmd.Stderr = os.Stderr
-	cmd.Run()
-	return nil
-}
-
-func getSSMEndpoint(region string) string {
-	return fmt.Sprintf("https://ssm.%s.amazonaws.com", region)
-}
